geometry fix

+++ template/geometry/area.cpp
@@ -1,4 +1,4 @@
-double area(polygon& poly) {
+ll area(polygon& poly) {
   ll tmp = 0;
   for(int i = 0;i < poly.size();i++) tmp += poly[i] / poly[(i + 1) % poly.size()];
   return abs(tmp);
